+++ packages/expo-router/CHANGELOG.md
@@ -252,6 +252,7 @@ _This version does not introduce any user-facing changes._
 
 ### 🐛 Bug fixes
 
+- Fix Jest 30 compatibility in testing library by removing internal `expect/build/matchers` import. ([#40184](https://github.com/expo/expo/issues/40184) by [@mvincentong](https://github.com/mvincentong))
 - Fix pinned `react-navigation` dependencies ([#43456](https://github.com/expo/expo/pull/43456) by [@kitten](https://github.com/kitten))
 
 ## 55.0.2 — 2026-02-25